Upgrade libfmt (#226)
* upgrade to libfmt 8.1.1
* fmt: add const to formatters
  new requirement according to fmt v8 changelog: https://github.com/fmtlib/fmt/releases/tag/8.0.0

* URI: add missing include
* Add static casts to std::errc because there is no default formatter
* BasicThreadPool example: adjust to new libfmt
* TimingCtx: move stream operator after formatter definition
* Utils_tests: adapt to changed vector fmt::formatter
  the formatter now uses `[]` instead of `{}` around vectors.

+++ src/serialiser/include/IoSerialiserJson.hpp
@@ -353,7 +353,7 @@ struct IoSerialiser<Json, T> {
             result = std::to_chars(data + start, data + size, value);
         }
         if (result.ec != std::errc()) {
-            throw ProtocolException("error({}) serialising number at buffer position: {}", result.ec, start);
+            throw ProtocolException("error({}) serialising number at buffer position: {}", static_cast<int>(result.ec), start);
         }
         buffer.resize(static_cast<size_t>(result.ptr - data)); // new position
     }
@@ -373,7 +373,7 @@ struct IoSerialiser<Json, T> {
         if constexpr (std::is_floating_point_v<T>) {
             const auto result = fast_float::from_chars(data + start, data + stop, value);
             if (result.ec != std::errc()) {
-                throw ProtocolException("error({}) parsing number at buffer position: {}", result.ec, start);
+                throw ProtocolException("error({}) parsing number at buffer position: {}", static_cast<int>(result.ec), start);
             }
             buffer.set_position(static_cast<size_t>(result.ptr - data)); // new position
             return;
@@ -381,7 +381,7 @@ struct IoSerialiser<Json, T> {
         // fall-back
         const auto result = std::from_chars(data + start, data + stop, value);
         if (result.ec != std::errc()) {
-            throw ProtocolException("error({}) parsing number at buffer position: {}", result.ec, start);
+            throw ProtocolException("error({}) parsing number at buffer position: {}", static_cast<int>(result.ec), start);
         }
         buffer.set_position(static_cast<size_t>(result.ptr - data)); // new position
     }
